app/src/main/java/mega/privacy/android/app/lollipop/controllers/ContactController.java
app/src/main/java/mega/privacy/android/app/lollipop/megaachievements/MegaReferralBonusesAdapter.java

+++ app/src/main/java/mega/privacy/android/app/lollipop/megaachievements/MegaReferralBonusesAdapter.java
@@ -191,47 +191,48 @@ public class MegaReferralBonusesAdapter extends RecyclerView.Adapter<MegaReferra
 		ReferralBonus referralBonus = (ReferralBonus) getItem(position);
 		holder.contactMail = referralBonus.getEmails().get(0);
 		MegaUser contact = megaApi.getContact(holder.contactMail);
-		long handle = contact.getHandle();
 
-		String fullName = getContactNameDB(megaApi, context, handle);
-		if(fullName ==  null) fullName = holder.contactMail;
+		if(contact != null)
+		{
+			long handle = contact.getHandle();
+			String fullName = getContactNameDB(megaApi, context, handle);
+			if(fullName ==  null) fullName = holder.contactMail;
+			logDebug("Contact: " + holder.contactMail + " name: " + fullName);
 
-		logDebug("Contact: " + holder.contactMail + " name: " + fullName);
+			holder.textViewContactName.setText(fullName);
 
-		holder.textViewContactName.setText(fullName);
+			holder.itemLayout.setBackgroundColor(Color.WHITE);
 
-		holder.itemLayout.setBackgroundColor(Color.WHITE);
+			Bitmap defaultAvatar = getDefaultAvatar(context, getColorAvatar(context, megaApi, contact), fullName, AVATAR_SIZE, true);
+			((ViewHolderReferralBonusesList)holder).imageView.setImageBitmap(defaultAvatar);
 
-		Bitmap defaultAvatar = getDefaultAvatar(context, getColorAvatar(context, megaApi, contact), fullName, AVATAR_SIZE, true);
-		((ViewHolderReferralBonusesList)holder).imageView.setImageBitmap(defaultAvatar);
+			UserAvatarListenerList listener = new UserAvatarListenerList(context, ((ViewHolderReferralBonusesList)holder), this);
 
-		UserAvatarListenerList listener = new UserAvatarListenerList(context, ((ViewHolderReferralBonusesList)holder), this);
-
-		File avatar = buildAvatarFile(context,holder.contactMail + ".jpg");
-        Bitmap bitmap = null;
-        if (isFileAvailable(avatar)){
-			if (avatar.length() > 0){
-				BitmapFactory.Options bOpts = new BitmapFactory.Options();
-				bOpts.inPurgeable = true;
-				bOpts.inInputShareable = true;
-				bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-				if (bitmap == null) {
-					avatar.delete();
-                    megaApi.getUserAvatar(contact,buildAvatarFile(context,contact.getEmail() + ".jpg").getAbsolutePath(),listener);
-                }
+			File avatar = buildAvatarFile(context,holder.contactMail + ".jpg");
+			Bitmap bitmap = null;
+			if (isFileAvailable(avatar)){
+				if (avatar.length() > 0){
+					BitmapFactory.Options bOpts = new BitmapFactory.Options();
+					bOpts.inPurgeable = true;
+					bOpts.inInputShareable = true;
+					bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+					if (bitmap == null) {
+						avatar.delete();
+						megaApi.getUserAvatar(holder.contactMail,buildAvatarFile(context,holder.contactMail + ".jpg").getAbsolutePath(),listener);
+					}
+					else{
+						logDebug("Do not ask for user avatar - its in cache: " + avatar.getAbsolutePath());
+						((ViewHolderReferralBonusesList)holder).imageView.setImageBitmap(bitmap);
+					}
+				}
 				else{
-					logDebug("Do not ask for user avatar - its in cache: " + avatar.getAbsolutePath());
-					((ViewHolderReferralBonusesList)holder).imageView.setImageBitmap(bitmap);
+					megaApi.getUserAvatar(holder.contactMail,buildAvatarFile(context,holder.contactMail + ".jpg").getAbsolutePath(),listener);
 				}
 			}
 			else{
-                megaApi.getUserAvatar(contact,buildAvatarFile(context,contact.getEmail() + ".jpg").getAbsolutePath(),listener);
+				megaApi.getUserAvatar(holder.contactMail,buildAvatarFile(context,holder.contactMail + ".jpg").getAbsolutePath(),listener);
 			}
 		}
-		else{
-            megaApi.getUserAvatar(contact,buildAvatarFile(context,contact.getEmail() + ".jpg").getAbsolutePath(),listener);
-		}
-
 
 		holder.textViewStorage.setText(getSizeString(referralBonus.getStorage()));
 		holder.textViewTransfer.setText(getSizeString(referralBonus.getTransfer()));
